Move `dict`-related functions to own module
