+++ pydian/dicts/core.py
@@ -1,8 +1,8 @@
 from collections.abc import Iterable
 from typing import Any
 
-from .lib.types import DROP, ApplyFunc, ConditionalCheck
-from .lib.util import _nested_get, _nested_set, flatten_list, get_tokenized_keypath
+from ..lib.types import DROP, ApplyFunc, ConditionalCheck
+from ..lib.util import _nested_get, flatten_list, get_tokenized_keypath
 from .mapper import _MapperContextStrict
