+++ pydian/dicts/mapper.py
@@ -2,8 +2,8 @@ import contextvars
 from contextlib import contextmanager
 from typing import Any
 
-from .lib.types import DROP, KEEP, MappingFunc
-from .lib.util import (
+from ..lib.types import DROP, KEEP, MappingFunc
+from ..lib.util import (
     drop_keys,
     get_keys_containing_class,
     impute_enum_values,
