+++ tests/dicts/test_core.py
